@@ -1,6 +1,8 @@
 module github.com/memoio/mefs-gateway
 
-go 1.16
+go 1.23.0
+
+toolchain go1.24.1
 
 require (
 	github.com/ipfs/go-blockservice v0.2.1
