@@ -7,15 +7,16 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	"path"
 	"strconv"
+	"strings"
 	"sync"
 	"time"
 
+	mtypes "github.com/memoio/go-mefs-v2/lib/types"
 	minio "github.com/memoio/minio/cmd"
 	"github.com/minio/cli"
 	"github.com/minio/madmin-go"
@@ -161,10 +162,10 @@ func (g *Mefs) NewGatewayLayer(creds madmin.Credentials) (minio.ObjectLayer, err
 	gw.useLocal = viper.GetBool("common.use_local")
 	gw.useS3 = viper.GetBool("common.use_s3")
 	gw.useMemo = viper.GetBool("common.use_memo")
+	gw.useMeeda = viper.GetBool("common.use_meeda")
 	gw.readOnly = viper.GetBool("common.read_only")
-	gw.repoDir = viper.GetString("common.repo_dir")
 
-	if !gw.useLocal && !gw.useS3 && !gw.useMemo {
+	if !gw.useLocal && !gw.useS3 && !gw.useMemo && !gw.useMeeda {
 		return nil, errors.New("must choose a backend")
 	}
 
@@ -207,7 +208,7 @@ func (g *Mefs) NewGatewayLayer(creds madmin.Credentials) (minio.ObjectLayer, err
 		}
 
 		// 扫描所有的Bucket进行判断
-		for i := 0; i < len(bs); i++ {
+		for i := range bs {
 			if bs[i].Name == BucketName {
 				break
 			}
@@ -229,21 +230,45 @@ func (g *Mefs) NewGatewayLayer(creds madmin.Credentials) (minio.ObjectLayer, err
 	}
 
 	if gw.useMemo {
-		var repoDir string
-		if gw.repoDir == "" {
+		repoDir := viper.GetString("memo.repo_dir")
+
+		if repoDir == "" {
 			repoDir = gw.rootpath
-		} else {
-			repoDir = gw.repoDir
 		}
 		gw.memofs, err = memo.NewMemofs(repoDir)
 		if err != nil {
 			return nil, err
+		}
+
+		// check bucket if exist
+		bi, err := gw.memofs.GetBucketInfo(context.Background(), BucketName)
+		if err != nil {
+			if !strings.Contains(err.Error(), "already exist") {
+				return nil, err
+			}
+			err = gw.memofs.MakeBucketWithLocation(context.Background(), BucketName)
+			if err != nil {
+				return nil, err
+			}
+		}
+
+		if !bi.Confirmed {
+			time.Sleep(1 * time.Minute)
 		}
 	}
 
 	if gw.useIpfs {
 		host := viper.GetString("common.ipfs_host")
 		gw.ipfs = memo.NewIpfsClient(host)
+	}
+
+	if gw.useMeeda {
+		meeda_repo := viper.GetString("meeda.repo_dir")
+		url := viper.GetString("meeda.url")
+		gw.meedafs, err = memo.NewMeedaFs(meeda_repo, url)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	// 读取已使用的空间大小
@@ -271,14 +296,17 @@ type lfsGateway struct {
 	useLocal bool
 	useMemo  bool
 	useIpfs  bool
+	useMeeda bool
 
 	readOnly bool
-	repoDir  string
+	// memoRepoDir  string
+	// meedaRepoDir string
 
 	usedBytes uint64
 
 	memofs  *memo.MemoFs
 	localfs *utils.LocalFS
+	meedafs *memo.MeedaFs
 	Client  *miniogo.Core
 	ipfs    *memo.IpFs
 	polices map[string]*policy.Policy
@@ -339,9 +367,20 @@ func (l *lfsGateway) SetBucketPolicy(ctx context.Context, bucket string, bucketP
 
 // GetBucketPolicy will get policy on bucket.
 func (l *lfsGateway) GetBucketPolicy(ctx context.Context, bucket string) (*policy.Policy, error) {
-	bi, err := l.memofs.GetBucketInfo(ctx, bucket)
-	if err != nil {
-		return nil, err
+	var bi mtypes.BucketInfo
+	var err error
+	if l.useMeeda {
+		bi, err = l.meedafs.GetBucketInfo(ctx)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	if l.useMemo {
+		bi, err = l.memofs.GetBucketInfo(ctx, bucket)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	pb, ok := l.polices[bucket]
@@ -450,7 +489,7 @@ func (l *lfsGateway) GetBucketInfo(ctx context.Context, bucket string) (bi minio
 		}
 	}
 
-	if l.useLocal {
+	if l.useLocal || l.useMeeda {
 		return minio.BucketInfo{
 			Name: bucket,
 		}, nil
@@ -510,6 +549,7 @@ func (l *lfsGateway) DeleteBucket(ctx context.Context, bucket string, opts minio
 
 // ListObjects lists all blobs in LFS bucket filtered by prefix.
 func (l *lfsGateway) ListObjects(ctx context.Context, bucket, prefix, marker, delimiter string, maxKeys int) (loi minio.ListObjectsInfo, err error) {
+	log.Println("ListObjects ", bucket)
 	if delimiter == SlashSeparator && prefix == SlashSeparator {
 		return loi, nil
 	}
@@ -545,6 +585,33 @@ func (l *lfsGateway) ListObjects(ctx context.Context, bucket, prefix, marker, de
 		return loi, nil
 	}
 
+	if l.useMeeda {
+		mloi, err := l.meedafs.ListObjects(ctx, prefix, marker, delimiter, maxKeys)
+		if err != nil {
+			return loi, err
+		}
+		ud := make(map[string]string)
+		ud["x-amz-meta-mode"] = "33204"
+		for _, oi := range mloi.Objects {
+			etag, _ := metag.ToString(oi.ETag)
+			ud["x-amz-meta-mtime"] = strconv.FormatInt(oi.GetTime(), 10)
+			loi.Objects = append(loi.Objects, minio.ObjectInfo{
+				Bucket:      bucket,
+				Name:        oi.GetName(),
+				ModTime:     time.Unix(oi.GetTime(), 0).UTC(),
+				Size:        int64(oi.Size),
+				ETag:        etag,
+				UserDefined: ud,
+			})
+		}
+
+		loi.IsTruncated = mloi.IsTruncated
+		loi.NextMarker = mloi.NextMarker
+		loi.Prefixes = mloi.Prefixes
+
+		return loi, nil
+	}
+
 	if l.useS3 {
 		result, err := l.Client.ListObjects(bucket, prefix, marker, delimiter, maxKeys)
 		if err != nil {
@@ -553,8 +620,12 @@ func (l *lfsGateway) ListObjects(ctx context.Context, bucket, prefix, marker, de
 
 		return minio.FromMinioClientListBucketResult(bucket, result), nil
 	}
-	return l.localfs.ListObjects(bucket)
 
+	if l.useLocal {
+		return l.localfs.ListObjects(bucket)
+	}
+
+	return loi, minio.ErrorRespToObjectError(err, bucket)
 }
 
 // ListObjectsV2 lists all blobs in LFS bucket filtered by prefix
@@ -563,13 +634,13 @@ func (l *lfsGateway) ListObjectsV2(ctx context.Context, bucket, prefix, continua
 	// if bucket != BucketName {
 	// 	return loiv2, minio.BucketNotFound{Bucket: bucket}
 	// }
-	// log.Println("ListObjectsV2 ", bucket)
+	log.Println("GetObjectNInfo ", bucket)
 	marker := continuationToken
 	if marker == "" {
 		marker = startAfter
 	}
 
-	if l.useMemo {
+	if l.useMemo || l.useMeeda {
 		loi, err := l.ListObjects(ctx, bucket, prefix, marker, delimiter, maxKeys)
 		if err != nil {
 			return loiv2, err
@@ -601,6 +672,7 @@ func (l *lfsGateway) ListObjectsV2(ctx context.Context, bucket, prefix, continua
 
 // GetObjectNInfo - returns object info and locked object ReadCloser
 func (l *lfsGateway) GetObjectNInfo(ctx context.Context, bucket, object string, rs *minio.HTTPRangeSpec, h http.Header, lockType minio.LockType, opts minio.ObjectOptions) (gr *minio.GetObjectReader, err error) {
+	log.Println("GetObjectNInfo ", object)
 	var objInfo minio.ObjectInfo
 	objInfo, err = l.GetObjectInfo(ctx, bucket, object, opts)
 	if err != nil {
@@ -642,6 +714,7 @@ func (e InvalidRange) Error() string {
 // startOffset indicates the starting read location of the object.
 // length indicates the total length of the object.
 func (l *lfsGateway) GetObject(ctx context.Context, bucketName, objectName string, startOffset, length int64, writer io.Writer, etag string, o minio.ObjectOptions) error {
+	log.Println("GetObject ", objectName)
 	if length < 0 && length != -1 {
 		return minio.ErrorRespToObjectError(minio.InvalidRange{}, bucketName, objectName)
 	}
@@ -652,6 +725,7 @@ func (l *lfsGateway) GetObject(ctx context.Context, bucketName, objectName strin
 
 	if l.useLocal {
 		object, size, err := l.localfs.GetObject(bucketName, objectName, startOffset)
+		fmt.Println("local object: ")
 		if err == nil {
 			defer object.Close()
 			reader := io.LimitReader(object, length)
@@ -714,6 +788,14 @@ func (l *lfsGateway) GetObject(ctx context.Context, bucketName, objectName strin
 		}
 		return nil
 	}
+
+	if l.useMeeda {
+		err := l.meedafs.GetObject(ctx, objectName, startOffset, length, writer)
+		if err != nil {
+			return err
+		}
+		return nil
+	}
 	if l.useS3 {
 		opts := miniogo.GetObjectOptions{}
 		opts.ServerSideEncryption = o.ServerSideEncryption
@@ -748,7 +830,7 @@ func (l *lfsGateway) GetObjectInfo(ctx context.Context, bucket, object string, o
 	// if bucket != BucketName {
 	// 	return objInfo, minio.BucketNotFound{Bucket: bucket}
 	// }
-	// log.Println("GetObjectInfo ", object)
+	log.Println("GetObjectInfo ", object)
 	if l.useIpfs {
 		objInfo.Size = -1
 		objInfo.Bucket = "nft"
@@ -787,22 +869,60 @@ func (l *lfsGateway) GetObjectInfo(ctx context.Context, bucket, object string, o
 		return minio.FromMinioClientObjectInfo(bucket, oi), nil
 	}
 
-	return l.localfs.GetObjectInfo(bucket, object)
+	if l.useMeeda {
+		moi, err := l.meedafs.GetObjectInfo(ctx, object)
+		if err != nil {
+			return objInfo, err
+		}
+		ud := make(map[string]string)
+		ud["x-amz-meta-mode"] = "33204"
+		ud["x-amz-meta-mtime"] = strconv.FormatInt(moi.GetTime(), 10)
+		// need handle ETag
+		etag, _ := metag.ToString(moi.ETag)
+		oi := miniogo.ObjectInfo{
+			Key:  moi.Name,
+			ETag: etag,
+			Size: int64(moi.Size),
+		}
+
+		return minio.FromMinioClientObjectInfo(bucket, oi), nil
+	}
+	if l.useLocal {
+		moi, err := l.localfs.GetObjectInfo(bucket, object)
+		if err != nil {
+			return objInfo, err
+		}
+		ud := make(map[string]string)
+		ud["x-amz-meta-mode"] = "33204"
+		ud["x-amz-meta-mtime"] = strconv.FormatInt(moi.ModTime.Unix(), 10)
+
+		oi := miniogo.ObjectInfo{
+			Key:  moi.Name,
+			ETag: moi.ETag,
+			Size: int64(moi.Size),
+		}
+
+		return minio.FromMinioClientObjectInfo(bucket, oi), nil
+	}
+
+	return objInfo, minio.ObjectNotFound{Bucket: bucket, Object: object}
 }
 
 // spilt reader to qiniu and memo
-func readerSpilt(reader io.Reader, reader1, reader2 *bytes.Buffer) error {
-	b, err := ioutil.ReadAll(reader)
+func readerSpilt(reader io.Reader, reader1, reader2 *bytes.Buffer) ([]byte, error) {
+	b, err := io.ReadAll(reader)
 	if err != nil {
 		log.Println("readerspilt err: ", err)
+		return nil, err
 	}
 	reader1.Write(b)
 	reader2.Write(b)
-	return err
+	return b, err
 }
 
 // PutObject creates a new object with the incoming data.
 func (l *lfsGateway) PutObject(ctx context.Context, bucket, object string, r *minio.PutObjReader, opts minio.ObjectOptions) (objInfo minio.ObjectInfo, err error) {
+	log.Println("PutObject ", object)
 	if l.readOnly {
 		return objInfo, minio.PrefixAccessDenied{Bucket: bucket}
 	}
@@ -839,157 +959,288 @@ func (l *lfsGateway) PutObject(ctx context.Context, bucket, object string, r *mi
 	}
 
 	var reader io.Reader = limitedReader
+	var moi minio.ObjectInfo
+
+	b, err := io.ReadAll(reader)
+	if err != nil {
+		return objInfo, err
+	}
+	if l.useMemo {
+		readerMemo := new(bytes.Buffer)
+		readerMemo.Write(b)
+		moi, err = l.putObjectMemo(ctx, bucket, object, readerMemo, opts)
+		if err != nil {
+			return objInfo, err
+		}
+		log.Println("put object Memo success")
+	}
+
+	if l.useLocal {
+		readerLocal := new(bytes.Buffer)
+		readerLocal.Write(b)
+		go func(reader io.Reader) {
+			_, err := l.putObjectLocal(ctx, bucket, object, reader, opts)
+			if err != nil {
+				log.Println("put object Local error:", err)
+			}
+			log.Println("put object Local success")
+		}(readerLocal)
+	}
+
+	if l.useMeeda {
+		go func(b1 []byte) {
+			_, err = l.putObjectMeeda(ctx, bucket, object, b1, opts)
+			if err != nil {
+				log.Println("put object Meeda error:", err)
+			}
+			log.Println("put object Meeda success")
+		}(b)
+	}
+
+	if l.useS3 {
+		readerS3 := new(bytes.Buffer)
+		readerS3.Write(b)
+		go func(reader io.Reader, size int64) {
+			_, err = l.putObjectS3(ctx, bucket, object, reader, size, putOpts)
+			if err != nil {
+				log.Println("put object S3 error:", err)
+			}
+			log.Println("put object S3 success")
+		}(readerS3, data.Size())
+
+	}
+
+	// 如果要存到本地
+	// if l.useLocal {
+	// 	moi, err = l.putObjectLocal(ctx, bucket, object, reader, opts)
+	// 	if err != nil {
+	// 		return objInfo, err
+	// 	}
+	// }
+
+	// if l.useS3 && l.useMemo {
+	// 	reader1 := new(bytes.Buffer)
+	// 	reader2 := new(bytes.Buffer)
+	// 	b, err := readerSpilt(reader, reader1, reader2)
+	// 	if err != nil {
+	// 		return objInfo, err
+	// 	}
+	// 	hashReader, err := minio.NewhashReader(reader1, data.Size(), "", "", data.Size())
+	// 	if err != nil {
+	// 		return objInfo, err
+	// 	}
+	// 	rawReader := hashReader
+	// 	pReader := minio.NewPutObjReader(rawReader)
+	// 	moi, err := l.memofs.PutObject(ctx, bucket, object, reader2, opts.UserDefined)
+	// 	if err != nil {
+	// 		return objInfo, err
+	// 	}
+	// 	etag, _ := metag.ToString(moi.ETag)
+	// 	oi = miniogo.ObjectInfo{
+	// 		ETag:     etag,
+	// 		Size:     int64(moi.Size),
+	// 		Key:      object,
+	// 		Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
+	// 	}
+
+	// 	go func(pr *minio.PutObjReader, size int64) {
+	// 		qBucketname := viper.GetString("common.bucketname")
+	// 		_, err := l.Client.PutObject(context.TODO(), qBucketname, object, pReader, size, "", "", putOpts)
+	// 		if err != nil {
+	// 			log.Println("put object error:", err)
+	// 		} else {
+	// 			log.Println("Success!")
+	// 		}
+	// 	}(pReader, data.Size())
+	// 	if l.useMeeda {
+	// 		go func(b1 []byte) {
+	// 			_, err := l.meedafs.Putobject(b1)
+	// 			if err != nil {
+	// 				log.Println("put object error:", err)
+	// 			} else {
+	// 				log.Println("Success!")
+	// 			}
+	// 		}(b)
+	// 	}
+
+	// 	if l.useLocal {
+	// 		closer.Close()
+	// 		err = l.localfs.FinishPut(bucket, object, oi.Size, true)
+	// 		if err != nil {
+	// 			log.Println("finish put error:", err)
+	// 		}
+	// 	}
+
+	// 	l.addUsedBytes(limitedReader.ReadedBytes())
+	// 	return minio.FromMinioClientObjectInfo(bucket, oi), nil
+	// }
+
+	// if l.useIpfs && l.useMemo {
+	// 	reader1 := new(bytes.Buffer)
+	// 	reader2 := new(bytes.Buffer)
+	// 	_, err = readerSpilt(reader, reader1, reader2)
+	// 	if err != nil {
+	// 		return objInfo, err
+	// 	}
+	// 	moi, err := l.memofs.PutObject(ctx, bucket, object, reader2, opts.UserDefined)
+	// 	if err != nil {
+	// 		return objInfo, err
+	// 	}
+	// 	etag, _ := metag.ToString(moi.ETag)
+	// 	oi = miniogo.ObjectInfo{
+	// 		ETag:     etag,
+	// 		Size:     int64(moi.Size),
+	// 		Key:      object,
+	// 		Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
+	// 	}
+
+	// 	go func(reader io.Reader) {
+	// 		_, err := l.ipfs.Putobject(reader)
+	// 		if err != nil {
+	// 			log.Println("put object error:", err)
+	// 		} else {
+	// 			log.Println("Success!")
+	// 		}
+	// 	}(reader1)
+
+	// 	l.addUsedBytes(limitedReader.ReadedBytes())
+	// 	return minio.FromMinioClientObjectInfo(bucket, oi), nil
+	// }
+	// if l.useMemo {
+	// 	moi, err := l.memofs.PutObject(ctx, bucket, object, reader, opts.UserDefined)
+	// 	if err != nil {
+	// 		return objInfo, err
+	// 	}
+	// 	etag, _ := metag.ToString(moi.ETag)
+	// 	oi = miniogo.ObjectInfo{
+	// 		ETag:     etag,
+	// 		Size:     int64(moi.Size),
+	// 		Key:      object,
+	// 		Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
+	// 	}
+	// 	l.addUsedBytes(limitedReader.ReadedBytes())
+	// 	return minio.FromMinioClientObjectInfo(bucket, oi), nil
+	// }
+
+	// if l.useS3 {
+	// 	ui, err := l.Client.PutObject(ctx, bucket, object, reader, data.Size(), data.MD5Base64String(), data.SHA256HexString(), putOpts)
+	// 	if err != nil {
+	// 		if l.useLocal {
+	// 			closer.Close()
+	// 			err = l.localfs.FinishPut(bucket, object, 0, false)
+	// 			if err != nil {
+	// 				return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
+	// 			}
+	// 		}
+	// 		return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
+	// 	}
+
+	// 	// On success, populate the key & metadata so they are present in the notification
+	// 	oi = miniogo.ObjectInfo{
+	// 		ETag:     ui.ETag,
+	// 		Size:     ui.Size,
+	// 		Key:      object,
+	// 		Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
+	// 	}
+	// } else {
+	// 	w := &utils.EmptyWriter{}
+	// 	_, err := io.Copy(w, reader)
+	// 	if err != nil {
+	// 		if l.useLocal {
+	// 			closer.Close()
+	// 			err = l.localfs.FinishPut(bucket, object, 0, false)
+	// 			if err != nil {
+	// 				return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
+	// 			}
+	// 		}
+
+	// 		return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
+	// 	}
+
+	// 	// On success, populate the key & metadata so they are present in the notification
+	// 	oi = miniogo.ObjectInfo{
+	// 		Size:     w.Size(),
+	// 		Key:      object,
+	// 		Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
+	// 	}
+	// }
+
+	// 记录新增的存储空间
+	l.addUsedBytes(limitedReader.ReadedBytes())
+
+	return moi, nil
+}
+
+func (l *lfsGateway) putObjectMemo(ctx context.Context, bucket, object string, reader io.Reader, opts minio.ObjectOptions) (objInfo minio.ObjectInfo, err error) {
+	moi, err := l.memofs.PutObject(ctx, bucket, object, reader, opts.UserDefined)
+	if err != nil {
+		return objInfo, err
+	}
+	etag, _ := metag.ToString(moi.ETag)
+	oi := miniogo.ObjectInfo{
+		ETag:     etag,
+		Size:     int64(moi.Size),
+		Key:      object,
+		Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
+	}
+
+	return minio.FromMinioClientObjectInfo(bucket, oi), nil
+}
+
+func (l *lfsGateway) putObjectS3(ctx context.Context, bucket, object string, reader io.Reader, size int64, putOpts miniogo.PutObjectOptions) (objInfo minio.ObjectInfo, err error) {
+	hashReader, err := minio.NewhashReader(reader, size, "", "", size)
+	if err != nil {
+		return objInfo, err
+	}
+	rawReader := hashReader
+	pReader := minio.NewPutObjReader(rawReader)
+
+	ui, err := l.Client.PutObject(ctx, bucket, object, pReader, size, "", "", putOpts)
+	if err != nil {
+		return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
+	}
+
+	// On success, populate the key & metadata so they are present in the notification
+	oi := miniogo.ObjectInfo{
+		ETag:     ui.ETag,
+		Size:     ui.Size,
+		Key:      object,
+		Metadata: minio.ToMinioClientObjectInfoMetadata(putOpts.UserMetadata),
+	}
+
+	return minio.FromMinioClientObjectInfo(bucket, oi), nil
+}
+
+func (l *lfsGateway) putObjectLocal(ctx context.Context, bucket, object string, reader io.Reader, opts minio.ObjectOptions) (objInfo minio.ObjectInfo, err error) {
 	var closer io.Closer
 
 	var oi miniogo.ObjectInfo
 
-	// 如果要存到本地
-	if l.useLocal {
-		reader, closer, err = l.localfs.PutObject(bucket, object, reader)
-		if err != nil {
-			return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
-		}
+	reader, closer, err = l.localfs.PutObject(bucket, object, reader)
+	if err != nil {
+		return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
 	}
 
-	if l.useS3 && l.useMemo {
-		reader1 := new(bytes.Buffer)
-		reader2 := new(bytes.Buffer)
-		err = readerSpilt(reader, reader1, reader2)
-		if err != nil {
-			return objInfo, err
-		}
-		hashReader, err := minio.NewhashReader(reader1, data.Size(), "", "", data.Size())
-		if err != nil {
-			return objInfo, err
-		}
-		rawReader := hashReader
-		pReader := minio.NewPutObjReader(rawReader)
-		moi, err := l.memofs.PutObject(ctx, bucket, object, reader2, opts.UserDefined)
-		if err != nil {
-			return objInfo, err
-		}
-		etag, _ := metag.ToString(moi.ETag)
-		oi = miniogo.ObjectInfo{
-			ETag:     etag,
-			Size:     int64(moi.Size),
-			Key:      object,
-			Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
-		}
-
-		go func(pr *minio.PutObjReader, size int64) {
-			qBucketname := viper.GetString("common.bucketname")
-			_, err := l.Client.PutObject(context.TODO(), qBucketname, object, pReader, size, "", "", putOpts)
-			if err != nil {
-				log.Println("put object error:", err)
-			} else {
-				log.Println("Success!")
-			}
-		}(pReader, data.Size())
-
-		l.addUsedBytes(limitedReader.ReadedBytes())
-		return minio.FromMinioClientObjectInfo(bucket, oi), nil
+	closer.Close()
+	err = l.localfs.FinishPut(bucket, object, oi.Size, true)
+	if err != nil {
+		return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
 	}
 
-	if l.useIpfs && l.useMemo {
-		reader1 := new(bytes.Buffer)
-		reader2 := new(bytes.Buffer)
-		err = readerSpilt(reader, reader1, reader2)
-		if err != nil {
-			return objInfo, err
-		}
-		moi, err := l.memofs.PutObject(ctx, bucket, object, reader2, opts.UserDefined)
-		if err != nil {
-			return objInfo, err
-		}
-		etag, _ := metag.ToString(moi.ETag)
-		oi = miniogo.ObjectInfo{
-			ETag:     etag,
-			Size:     int64(moi.Size),
-			Key:      object,
-			Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
-		}
+	return minio.FromMinioClientObjectInfo(bucket, oi), nil
+}
 
-		go func(reader io.Reader) {
-			_, err := l.ipfs.Putobject(reader)
-			if err != nil {
-				log.Println("put object error:", err)
-			} else {
-				log.Println("Success!")
-			}
-		}(reader1)
-
-		l.addUsedBytes(limitedReader.ReadedBytes())
-		return minio.FromMinioClientObjectInfo(bucket, oi), nil
+func (l *lfsGateway) putObjectMeeda(ctx context.Context, bucket, object string, data []byte, opts minio.ObjectOptions) (objInfo minio.ObjectInfo, err error) {
+	id, err := l.meedafs.Putobject(data)
+	if err != nil {
+		return objInfo, err
 	}
-	if l.useMemo {
-		moi, err := l.memofs.PutObject(ctx, bucket, object, reader, opts.UserDefined)
-		if err != nil {
-			return objInfo, err
-		}
-		etag, _ := metag.ToString(moi.ETag)
-		oi = miniogo.ObjectInfo{
-			ETag:     etag,
-			Size:     int64(moi.Size),
-			Key:      object,
-			Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
-		}
-		l.addUsedBytes(limitedReader.ReadedBytes())
-		return minio.FromMinioClientObjectInfo(bucket, oi), nil
+	oi := miniogo.ObjectInfo{
+		ETag:     id,
+		Size:     int64(len(data)),
+		Key:      object,
+		Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
 	}
-
-	if l.useS3 {
-		ui, err := l.Client.PutObject(ctx, bucket, object, reader, data.Size(), data.MD5Base64String(), data.SHA256HexString(), putOpts)
-		if err != nil {
-			if l.useLocal {
-				closer.Close()
-				err = l.localfs.FinishPut(bucket, object, 0, false)
-				if err != nil {
-					return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
-				}
-			}
-			return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
-		}
-
-		// On success, populate the key & metadata so they are present in the notification
-		oi = miniogo.ObjectInfo{
-			ETag:     ui.ETag,
-			Size:     ui.Size,
-			Key:      object,
-			Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
-		}
-	} else {
-		w := &utils.EmptyWriter{}
-		_, err := io.Copy(w, reader)
-		if err != nil {
-			if l.useLocal {
-				closer.Close()
-				err = l.localfs.FinishPut(bucket, object, 0, false)
-				if err != nil {
-					return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
-				}
-			}
-
-			return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
-		}
-
-		// On success, populate the key & metadata so they are present in the notification
-		oi = miniogo.ObjectInfo{
-			Size:     w.Size(),
-			Key:      object,
-			Metadata: minio.ToMinioClientObjectInfoMetadata(opts.UserDefined),
-		}
-	}
-
-	if l.useLocal {
-		closer.Close()
-		err = l.localfs.FinishPut(bucket, object, oi.Size, true)
-		if err != nil {
-			return objInfo, minio.ErrorRespToObjectError(err, bucket, object)
-		}
-	}
-
-	// 记录新增的存储空间
-	l.addUsedBytes(limitedReader.ReadedBytes())
 
 	return minio.FromMinioClientObjectInfo(bucket, oi), nil
 }
